Frontend: Add authentication guard and automatic login check.
Before navigating to any page, whether the user is logged in is checked. If the user is logged in the store, everything's good. Otherwise, tries to log in with the backend.
Then, depending on whether the user is logged in or not, the navigation redirects to login or continues to the protected page.